Provide in memory account service
