+++ app/controllers/OERWorldMap.java
@@ -15,6 +15,8 @@ import play.i18n.Lang;
 import play.mvc.Controller;
 import play.mvc.With;
 import services.AccountService;
+import services.KeycloakAccountService;
+import services.MemoryAccountService;
 import services.QueryContext;
 import services.repository.BaseRepository;
 import services.repository.ElasticsearchRepository;
@@ -61,14 +63,18 @@ public abstract class OERWorldMap extends Controller {
   private static synchronized void createAccountService(Configuration aConf) {
     Configuration keycloakConfig = aConf.getConfig("keycloak");
     if (mAccountService == null) {
-      mAccountService = new AccountService(
-        keycloakConfig.getString("serverUrl"),
-        keycloakConfig.getString("realm"),
-        keycloakConfig.getString("username"),
-        keycloakConfig.getString("password"),
-        keycloakConfig.getString("client"),
-        new File(aConf.getString("ht.permissions")));
-      mAccountService.setApache2Ctl(aConf.getString("ht.apache2ctl.restart"));
+      if (keycloakConfig != null) {
+        mAccountService = new KeycloakAccountService(
+          keycloakConfig.getString("serverUrl"),
+          keycloakConfig.getString("realm"),
+          keycloakConfig.getString("username"),
+          keycloakConfig.getString("password"),
+          keycloakConfig.getString("client"),
+          new File(aConf.getString("ht.permissions")));
+        mAccountService.setApache2Ctl(aConf.getString("ht.apache2ctl.restart"));
+      } else {
+        mAccountService = new MemoryAccountService();
+      }
     }
   }
